+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.7...HEAD)
 
+### Fixed
+- Restrictions: add tests
+
 
 ## [v0.7.7 - 2024-11-08](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.6...v0.7.7)